代码优化
