+++ streamx-common/src/main/scala/com/streamxhub/streamx/common/enums/RestartStrategy.scala
@@ -18,9 +18,24 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.streamxhub.streamx.flink.core.scala.enums
+package com.streamxhub.streamx.common.enums
 
-object PlannerType extends Enumeration {
-  type PlannerType = Value
-  val blink, old, any = Value
+object RestartStrategy extends Enumeration {
+  type RestartStrategy = Value
+  val `fixed-delay`, `failure-rate`, `none` = Value
+
+  /**
+   *
+   * @param name
+   * @return
+   */
+  def byName(name: String): Value = {
+    if (name == null) null else {
+      values.find(_.toString.replace("$minus", "-").equalsIgnoreCase(name)) match {
+        case Some(v) => v
+        case _ => throw new IllegalArgumentException("[StreamX] RestartStrategy muse be (fixed-delay|failure-rate|none)")
+      }
+    }
+  }
+
 }
