+++ streamx-flink/streamx-flink-core/src/main/scala/com/streamxhub/streamx/flink/core/scala/source/MySQLSource.scala
@@ -20,11 +20,11 @@
  */
 package com.streamxhub.streamx.flink.core.scala.source
 
+import com.streamxhub.streamx.common.enums.ApiType
 import com.streamxhub.streamx.common.util.{JdbcUtils, Logger, Utils}
 import com.streamxhub.streamx.flink.core.java.function.{SQLQueryFunction, SQLResultFunction}
 import com.streamxhub.streamx.flink.core.scala.StreamingContext
-import com.streamxhub.streamx.flink.core.scala.enums.ApiType
-import com.streamxhub.streamx.flink.core.scala.enums.ApiType.ApiType
+import ApiType.ApiType
 import com.streamxhub.streamx.flink.core.scala.util.FlinkUtils
 import org.apache.flink.api.common.state.ListState
 import org.apache.flink.api.common.typeinfo.TypeInformation
